+++ Tests/ReindexerNet.EmbeddedTest/ReindexerBindingTest.cs
@@ -129,6 +129,8 @@ public class ReindexerBindingTest
     private const string FrameworkName = "NET7";
 #elif NET8_0
     private const string FrameworkName = "NET8";
+#elif NET9_0
+    private const string FrameworkName = "NET9";
 #elif NETCOREAPP2_2
     private const string FrameworkName="NETCOREAPP2_2";
 #elif NETCOREAPP3_1