+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.10] - 2024-11-28
+## Added
+- .NET 9.0 target
+- Added more tests for nullable types(sparse indicies)
+- Symbol packages
+### Fixed
+- Fixed Join queries for embedded client on .net 4.7.2
+
 ## [0.4.9 Embedded] - 2024-11-05
 ### Fixed
 - Face and swagger ui hadn't included in previous embedded windows package.